Replace uintptr_t with uint32_t for addresses
It seems all occurences with the wrong (non) sign extension during the address
calculation are now fixed and the uintptr_t isn't needed anymore.

+++ execute.h
@@ -56,7 +56,7 @@ EX:
             switch (op)
             {
                 signed int offset;
-                register uintptr_t addr;
+                register uint32_t addr;
 
                 case 000: /* SPECIAL */
                     switch (inst % 64)
